Add files via upload
The card credit validator will take every digit out of a 16-digit number inserted by the user and multiply it by 2. Then if the result of the multiplication is a 2-digit number, it will subtract 9.
After this, the sum of these multiplications must be a number divisible by 10. In this case, the card credit number is validated, else it is not a valid number.

Example: 2323 2323 2323 2323
After multiplication by 2: 4646 4646 4646 4646
Sum each digit: 4+6+4+6+4+6+4+6+4+6+4+6+4+6+4+6 = 80
80 is divisible by 10, so it is a valid number.